% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{betaDiff}
\alias{betaDiff}
\title{Simulate a diffusion process with beta invariant distribution}
\usage{
betaDiff(T = 10, dt = 0.01, alpha = 3, beta = 5, theta = 1)
}
\arguments{
\item{dt}{Equal subintervals}

\item{alpha}{The beta distribution parameter alpha}

\item{beta}{The beta distribution parameter beta}

\item{theta}{The beta distribution parameter theta}
}
\value{
A simulation of diffusion process given beta invariant distribution
}
\description{
Simulate a diffusion process with beta invariant distribution
}
\examples{
betaDiff()
betaDiff(alpha=2,beta=4,theta=0.5)
}
