% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.R
\name{betaEstimating}
\alias{betaEstimating}
\title{Title}
\usage{
betaEstimating(X, dt = 0.01)
}
\arguments{
\item{X}{A diffusion process with beta invariant distribution}

\item{dt}{Equal subintervals}
}
\value{
Esimates of parameters in diffusion model with beta invariant distribution
}
\description{
Title
}
